refactor: move Button component to shared directory
