+++ src/components/@shared/Button/index.jsx
@@ -1,4 +1,4 @@
-import Styled from 'components/Button/index.style';
+import Styled from 'components/@shared/Button/index.style';
 
 const Button = ({ children, ...rest }) => {
   return <Styled.Button {...rest}>{children}</Styled.Button>;
