+++ src/components/index.js
@@ -1,5 +1,5 @@
 export { default as Image } from 'components/Image';
 export { default as CartIcon } from 'components/CartIcon';
 export { default as QuantityController } from 'components/QuantityController';
-export { default as Button } from 'components/Button';
+export { default as Button } from 'components/@shared/Button';
 export { default as Header } from 'components/Header';
